add disableDetails. Moved Components in the proper directory
