+++ web/client/localConfig.json
@@ -4,26 +4,6 @@
         "useCORS": ["http://demo.geo-solutions.it/geoserver", "https://demo.geo-solutions.it:443/geoserver", "https://demo.geo-solutions.it/geoserver", "https://nominatim.openstreetmap.org", "http://cloudsdi.geo-solutions.it"]
     },
     "geoStoreUrl": "rest/geostore/",
-
-    "//":  "Genova",
-    "geoFencePath1": "geofence/rest",
-    "geoFenceUrl1": "https://mappe.comune.genova.it/",
-    "geoFenceServiceType1": "geofence",
-    "geoFenceLayerServiceType1": "csw",
-    "geoFenceGeoServerInstance1": {
-      "url": "https://mappe.comune.genova.it/geoserver/",
-      "id": 1
-    },
-   " // local": "",
-    "geoFencePath": "rest/geofence",
-    "geoFenceUrl": "/geoserver/",
-    "geoFenceServiceType": "geoserver",
-    "geoFenceLayerServiceType": "rest",
-    "geoFenceGeoServerInstance": {
-      "url": "/geoserver/",
-      "id": 1
-    },
-
     "printUrl": "pdf/info.json",
     "bingApiKey": "AhuXBu7ipR1gNbBfXhtUAyCZ6rkC5PkWpxs2MnMRZ1ZupxQfivjLCch22ozKSCAn",
     "mapquestApiKey": "__API_KEY_MAPQUEST__",
@@ -112,26 +92,6 @@
               "code": "es-ES",
               "description": "Español"
             }
-          }
-        },
-        "rulesmanager": {
-          "services": {
-            "WFS": [
-              "DescribeFeatureType",
-              "GetCapabilities",
-              "GetFeature",
-              "GetFeatureWithLock",
-              "LockFeature",
-              "Transaction"
-            ],
-            "WMS": [
-              "DescribeLayer",
-              "GetCapabilities",
-              "GetFeatureInfo",
-              "GetLegendGraphic",
-              "GetMap",
-              "GetStyles"
-            ]
           }
         }
       }
@@ -591,22 +551,6 @@
                 "containerPosition": "header"
             }
         }],
-        "manager": ["Header", "Redirect", "Manager", "Home", "UserManager", "GroupManager", "Footer"],
-        "rulesmanager": [
-            "Home",
-            "Redirect" ,
-            {
-                "name": "OmniBar",
-                    "cfg": {
-                        "containerPosition": "header",
-                        "className": "navbar shadow navbar-home"
-                    }
-            }, "ManagerMenu", "Login", "Language", "NavMenu",
-            "Attribution", "RulesDataGrid", "Notifications"
-            ,{
-                 "name": "RulesEditor",
-               "cfg": {"containerPosition": "columns"}
-            }
-        ]
+        "manager": ["Header", "Redirect", "Manager", "Home", "UserManager", "GroupManager", "Footer"]
     }
 }
